add product operation

+++ Controllers/productController.cs
@@ -1,7 +1,11 @@
-﻿using MCV_Empity.Models;
+﻿using AutoMapper;
+using MCV_Empity.Models;
 using MCV_Empity.Services.Implementations;
 using MCV_Empity.Services.InterFaces;
+using MCV_Empity.ViewModels;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.VisualBasic;
 using System.Reflection;
 
 namespace MCV_Empity.Controllers
@@ -10,26 +14,31 @@ namespace MCV_Empity.Controllers
 	{
 		private readonly IProductService _IproductService;
 		private IFileServiece _IFileServiece;
-		public productController(IProductService productService,IFileServiece fileServiece)
+		private ICategoryServices _categoryServices;
+		private readonly IMapper _mapper;
+		public productController(IProductService productService,IFileServiece fileServiece, ICategoryServices categoryServices ,IMapper mapper)
 		{
 			_IproductService = productService;
 			_IFileServiece = fileServiece;
+			_categoryServices = categoryServices;
+			_mapper=mapper;
 		}
-		public ActionResult Index()
+		public async Task< ActionResult> Index()
 		{
 
-			var products = _IproductService.GetProducts();
+			var products =await _IproductService.GetProducts();
 
 			return View(products);
 
 		}
-		public IActionResult Create()
+		public async Task<IActionResult> Create()
 		{
+			ViewBag.Category= new SelectList(await	_categoryServices.GetCategories(),"Id","Name");
 			return View();
 		}
 
 		[HttpPost]
-		public async Task <IActionResult> Create(product model)
+		public async Task <IActionResult> Create(AddProductViewModels model)
 		{
             try
             {
@@ -37,25 +46,20 @@ namespace MCV_Empity.Controllers
 
                 if (ModelState.IsValid)
                 {
-					var path = "";
-					if (model.File?.Length > 0)
-					{
-
-                       path = await _IFileServiece.Upload(model.File,"/image/");
-                        if (path == "Not saved")
-                        {
-                            return BadRequest();
-                        }
-                    }
-
-					model.path = path;
-
-
-				_IproductService.AddProduct(model);
-                return RedirectToAction(nameof(Index));
+					var product = _mapper.Map<product>(model);
+					//var product = new product()
+					//{
+					//	Name = model.Name,
+					//	Price = model.Price
+					//	,
+					//	CategoryId = model.CategoryId
+					//};
+				await _IproductService.AddProduct(product,model.Files);
+					return RedirectToAction(nameof(Index));
 				}
 				else
 				{
+					ViewBag.Category = new SelectList(await _categoryServices.GetCategories(), "Id", "Name");
 					return View(model);
 				}
 		
@@ -64,20 +68,21 @@ namespace MCV_Empity.Controllers
 			}
 			catch(Exception)
 			{
+				ViewBag.Category = new SelectList(await _categoryServices.GetCategories(), "Id", "Name");
 				return View();
 
 			}
 		
 		}
 
-		public IActionResult Details(int Id)
+		public  async Task< IActionResult> Details(int Id)
 		{
-			var products = _IproductService.GetProductById(Id);
+			var products =await _IproductService.GetProductById(Id);
 			return View(products);
 		}
 
-		public IActionResult Update(int Id) {
-			var pro = _IproductService.GetProductById(Id);
+		public async Task<IActionResult> UpdateAsync(int Id) {
+			var pro = await _IproductService.GetProductById(Id);
 			return View(pro);
 		}
 		[HttpPost]
@@ -87,7 +92,12 @@ namespace MCV_Empity.Controllers
 			{
 				if (ModelState.IsValid)
 				{
-					var path = model.path;
+					var product = await _IproductService.GetProductById(model.Id);
+					if(product == null)
+					{
+						return NotFound();
+					}
+					/*var path = model.path;
 					if (model.File?.Length > 0)
 					{
 						_IFileServiece.DeleteSource(path);
@@ -98,8 +108,8 @@ namespace MCV_Empity.Controllers
 						}
 					}
 
-					model.path = path;
-					_IproductService.UpdateProduct(model);
+					model.path = path;*/
+					await _IproductService.UpdateProduct(model);
 					return RedirectToAction(nameof(Index));
 				}
 				else
@@ -120,17 +130,38 @@ namespace MCV_Empity.Controllers
 
 		}
 
-		public IActionResult Delete(int Id)
+		public async Task<IActionResult> Delete(int Id)
 		{
-			var pro = _IproductService.GetProductById(Id);
+			var pro = await _IproductService.GetProductById(Id);
 			return View(pro);
 		}
 		[HttpPost]
-		public IActionResult IDelete(int Id)
+		public async Task< IActionResult> IDelete(int Id)
 		{
-			_IproductService.DeleteProduct(Id);
+			try
+			{
+				var productR = await _IproductService.GetProductById(Id);
+				if (productR == null)
+				{
+					return NotFound();
+				}
+				await _IproductService.DeleteProduct(productR);
+				return RedirectToAction(nameof(Index));
+			}
+			catch
+			{
+				return View();
+				
+			}
 
-			return RedirectToAction(nameof(Index));
+		}
+
+		[HttpPost]
+		public async Task< IActionResult >IsProductNameExist(string name)
+		{
+			var result =await _IproductService.IsProductNameExist(name);
+			
+			return Json(!result);
 		}
 	}
 }
